convert-platform: copy initramfs image too rootfs if existing
Tradional Linux deskop oriented distrobutions normally boot a
initramfs image, which prepares the hardware before switching to
the actual Linux kernel and filesystem.

This initramfs image is normally present on the boot partition,
together with the Linux kernel. To support updates of these
using Mender they must be part of the root filesystem.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>