extract 'boot-gap' and re-use that
'boot-gap' is the area starting at 0 and to the first partition. This
area contains the MBR, in some cases together with the bootloader.

To avoid re-building custom bootloaders lets try to re-use this area,
meaning that we also re-use the MBR but slighty modify instead of
creating a new from scratch and dropping the necessary bootstrap code
needed in case of a bootloader is present.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>