re-size root filesystem image based on a default of 4GB
Usually the images that we convert have been booted on to a device,
and normally the root filesystem has been expanded to occupy the
reminder of the disk (in range of GBs).

So far we have re-used the size from the input image when we generate
the output image as well, producing very large root filesystems
which are mainly empty.

This change will enforce a 4GB size on the root filesystem with no
regard to input size. The script will fail if it does not fit.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>