Dockerfile: use a non-hard-float toolchain
It seems that the Linaro hard-float toolchain does not produce
valid binaries for rpi0w even though it does support hard-float.
For this reason we change to a non-hard-float toolchain.

This does also effect e.g other boards which compile binaries
as they will no longer be linked with hard-float support. But
I do see this as a big problem as the things we compile mostl
likely do not utilize hard-float anyway, e.g GRUB and go do not
rely on hard-float optimization.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>