split up the docker files to support multiple architectures
Our current assumption has been that we are targeting older arm
cores that are running in 32-bit mode. This assumption prohibits
adding new board support which is not an arm 32-bit, e.g x86
or arm64.

To get around this, lets keep seperate Dockerfiles for different
archictures for now.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>