do not assume that /etc/hostname exists in image
If MENDER_DEVICE_TYPE is not set we try to get the device name
from /etc/hostname, but have seen images where this file does not
exist, e.g images generated with mkosi.

If /etc/hostname does not exist, fallback to a default string.

Changelog: Do not fail if /etc/hostname does not exist

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>