modify: add --keep-dirlinks option to rsync commands
Extract from rsync man page:

-K, --keep-dirlinks
      This  option  causes  the  receiving side  to  treat  a
      symlink  to  a  directory  as though  it  were  a  real
      directory, but only if it matches a real directory from
      the  sender.   Without   this  option,  the  receiver’s
      symlink  would  be deleted  and  replaced  with a  real
      directory.

This means if e.g /lib is a symlink in rootfs and we try to rsync
a directory which contains a /lib directory, without the 'K' option
the symlink would be overwritten by the directory in the src, and
this would probably break the system.

Distrobutions which have applied the "usr merge" [1] approach, will
have /lib as a symlink to /usr/lib.

[1]. https://www.freedesktop.org/wiki/Software/systemd/TheCaseForTheUsrMerge/

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>