configs/raspberrypi: add symlink to cmdline.txt and overlays directory
Certain applications on Raspbian will try to look these files up to
see what configurations are set to determinate functionality. Lets
try to keep compatability as much as possible of a converted image.

Though even with this change we are not 100 % compatible and we now
that tools like 'raspi-config' will not work on a converted image
due to us changing the mount point of the boot partition from /boot
to /uboot, but this is a configuration tool that you should not
run on a converted image anyway so we can live with breaking this.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>