package: Support bmap index file generation upon request.
Bmaptool is a generic tool for creating the block map (bmap) for a file and
copying files using the block map. The idea is that large files, like raw
system image files, can be copied or flashed a lot faster and more reliably
with bmaptool than with traditional tools, like "dd" or "cp".

Signed-off-by: Drew Moseley <drew.moseley@northern.tech>

Tweaked the 'bmaptool' check not to silently fail and added links to
bmap-tools repository.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>